Make ShaderVariation class header API-agnostic.
