+++ Source/Urho3D/Graphics/GraphicsImpl.h
@@ -22,6 +22,9 @@
 
 #pragma once
 
+// Note: GraphicsImpl class is purposefully API-specific. It should not be used by Urho3D client applications,
+// unless required for e.g. integration of 3rd party libraries that interface directly with the graphics device.
+
 #if defined(URHO3D_OPENGL)
 #include "OpenGL/OGLGraphicsImpl.h"
 #elif defined(URHO3D_D3D11)
