+++ Source/Urho3D/Graphics/View.cpp
@@ -603,11 +603,8 @@ void View::Render()
 
     // It is possible, though not recommended, that the same camera is used for multiple main views. Set automatic aspect ratio
     // to ensure correct projection will be used
-    if (camera_)
-    {
-        if (camera_->GetAutoAspectRatio())
-            camera_->SetAspectRatioInternal((float)(viewSize_.x_) / (float)(viewSize_.y_));
-    }
+    if (camera_ && camera_->GetAutoAspectRatio())
+        camera_->SetAspectRatioInternal((float)(viewSize_.x_) / (float)(viewSize_.y_));
 
     // Bind the face selection and indirection cube maps for point light shadows
 #ifndef GL_ES_VERSION_2_0
